Added and reorganized starter files for Signup Use case
