+++ src/app/Main.java
@@ -1,11 +1,11 @@
 package src.app;
 
-import interface_adapter.LoginViewModel;
-import interface_adapter.SignupViewModel;
-import interface_adapter.ViewManagerModel;
-import view.LoginView;
-import view.SignupView;
-import view.ViewManager;
+import src.interface_adapter.login.LoginViewModel;
+import src.interface_adapter.signup.SignupViewModel;
+import src.interface_adapter.login.ViewManagerModel;
+import src.view.LoginView;
+import src.view.SignupView;
+import src.view.ViewManager;
 
 
 public class Main {
