+++ src/interface_adapter/ViewModel.java
@@ -1,2 +1,17 @@
-package src.interface_adapter;public class ViewModel {
+package src.interface_adapter;
+
+import java.beans.PropertyChangeListener;
+
+public abstract class ViewModel {
+    private String viewName;
+
+    public ViewModel(String viewName) {
+        this.viewName = viewName;
+    }
+    public String getViewName() {
+        return this.viewName;
+    }
+
+    public abstract void firePropertyChanged();
+    public abstract void addPropertyChangeListener(PropertyChangeListener listener);
 }
