+++ src/interface_adapter/signup/SignupController.java
@@ -1,2 +1,8 @@
-package src.interface_adapter;public class SignupController {
+package src.interface_adapter.signup;
+
+import src.use_case.signup.SignupInputBoundary;
+import src.use_case.signup.SignupInputData
+
+public class SignupController {
+
 }
