+++ src/interface_adapter/signup/SignupPresenter.java
@@ -1,2 +1,12 @@
-package src.interface_adapter;public class SignupPresenter {
+package src.interface_adapter.signup;
+
+import src.use_case.signup.SignupOutputBoundary;
+import src.use_case.signup.SignupOutputData;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class SignupPresenter implements SignupOutputBoundary {
+
 }
+
