+++ src/interface_adapter/signup/SignupState.java
@@ -1,2 +1,4 @@
-package src.interface_adapter;public class SignupState {
+package src.interface_adapter.signup;
+
+public class SignupState {
 }
