+++ src/interface_adapter/signup/SignupViewModel.java
@@ -1,2 +1,10 @@
-package src.interface_adapter;public class SignupViewModel {
+package src.interface_adapter.signup;
+
+import src.interface_adapter.ViewModel;
+
+import java.beans.PropertyChangeListener;
+import java.beans.PropertyChangeSupport;
+
+public class SignupViewModel extends ViewModel {
+
 }
