+++ src/use_case/nutrients/NutrientsInteractor.java
@@ -1,2 +1,4 @@
-package src.use_case.nutrients;public class NutrientsInteractor {
+package src.use_case.nutrients;
+
+public class NutrientsInteractor {
 }
