+++ src/use_case/nutrients/NutrientsOutputBoundary.java
@@ -1,2 +1,5 @@
-package src.use_case.nutrients;public interface NutrientsOutputBoundary {
+package src.use_case.nutrients;
+
+public interface NutrientsOutputBoundary {
+
 }
