+++ src/use_case/nutrients/NutrientsUserDataAccessInterface.java
@@ -1,2 +1,4 @@
-package src.use_case.nutrients;public interface NutrientsUserDataAccessInterface {
+package src.use_case.nutrients;
+
+public interface NutrientsUserDataAccessInterface {
 }
