+++ src/use_case/signup/SignupInputData.java
@@ -1,6 +1,4 @@
 package src.use_case.signup;
 
 public class SignupInputData {
-
-
 }
